app/requirements.txt
discord.py
python-dotenv
pandas
numpy
matplotlib
watchfiles
gspread
dotenv